requirements.txt
streamlit
requests
google-genai
duckduckgo-search
pydantic
python-dotenv